Fix default values of flags arguments
The previous default value 'false' of flags argument
(action='store_true') would actually evaluate to True in a boolean
evaluation context (e.g. 'bool(value)' or 'if value:'). Hopefully, we
used 'if <value> is True:' so this did not pose a problem before. This
is also changed as 'if <value>:' for consistency.