Add a new "Work in progress" ui module with a main
This module contains a "Work In Progress" main() which plugs in new UI
components fed with read data. The module has no test at all, these will
be added later on.

We also add a --ng option to select the new UI from pg_activity program.
As this is experimental, we also add a 'z' key to exit the program in a
debugger.