Define Data class an attr-class and clean it up
In addition to using attrs, we turn pg_connect() method into a
classmethod that is then used as a builder from utils.pg_connect() (and
used by ui module).

The refresh_dbsize attribute is no long used as it's not handled yet in
new UI.

We finally remove now useless attributes of Data class.