Improve data model w.r.t. local/non-local mode
We introduce an ActivityStats type that models possible combination of
activities type (running/waiting/blocking) and system information only
in local mode. This way, we enforce consistent data at type level and
make it impossible to pass a non-None SystemInfo value when not in local
mode.

Sorting activities is now done within screen() view to limit repetition
in ui.main().