Update installation instructions
We now only document installation with pip and drop instructions using
sudo, which are not recommended unless people know what they are doing.

We also drop mention of the --with-man option, which only works for
system-wide installation.

The 'Dependencies' section is outdated and no longer needed when
installing using pip.

We also document Debian installation as an alternative.