Only pass non-None changes to ui.evolve()
In particular, we now compute the new sort_key value based on the
updated query_mode, meaning that if we received, e.g. <F2>, we'd reset
the sort_key accordingly.

This fixes a crash when navigating between running queries mode and
other ones while being sorting on, e.g., CPU.