Do not change UI state while in pause
When in pause mode, we do not query postgres for new processes so that
if we change the query mode in the meantime, the processes list would be
unrelated to the UI state. To avoid this, we prevent UI.evolve() to act
if the UI object is in pause. For clarity, we change the in_pause flag
with a new toggle_pause() method.

A regression test is added.