Display pg_stat_activity's wait_event when available
In the 'wait' column, we now display the 'wait_event' column of
pg_stat_activity when it's available (i.e. from postgresql version 10)
instead of a boolean value (which previously was kind of arbitrary,
since only a subset of waiting events would be considered). For older
postgresql versions, we still use the 'waiting' column as a Boolean
(Y/N) value. Previous coloring still applies (green for non-waiting, red
for waiting).