Handle scrolling
When the line count overflows outside the screen, we can now properly
scroll down. If the focus reaches 4/5 of the available lines on screen,
the first line is removed and we display a new line at the bottom. When
we reach the last line, pressing down will focus on the fist line again.
Likewise, if the up key is pressed when the focus is on the first line,
we focus on the last one.