Fix scrolling when wrapping is involved
With a wrapped display mode the current display is bugged. If some
queries are wrapped before the line we put the focus on. It's possible
that the said line is out of the screen.

With this patch, the line with the focus is the always the first one.

When a the focus is on a selected line, the text is now bold with a
lighter brand of cyan. It is otherwise difficult to see if the first
ligne of the display is selected since it always has the focus.