Rework types.Column model
We drop the "template_h" field and instead use "min_width", "max_width"
and "justify" fields which are easier to understand.

Table columns in columns_header() and processes_rows() are rendered in a
more natural way: just values interspersed with a white space.

This produces a bit of churn in test data but things are mostly the same
modulo a white white-space changes.