Add support for using Psycopg database driver
In addition to using psycopg2, we now support psycopg (3) as a database
driver. The "pg" module is extended to first try psycopg and eventually
fall back to using psycopg2 if the former is not available.

Due to the subtle differences between psycopg2 and psycopg 3, we
introduce a few compatibility functions.

For testing, we can force the use of psycopg2 by setting
_PGACTIVITY_USE_PSYCOPG2 environment variable. This is done in tox,
through the new -psycopg2 variant of pyX.Y environments and populated in
github workflow.

We add two setuptools "extras": psycopg2 and psycopg in order make
installation easier.