Compose 'CASE WHEN <condition>...' clauses client-side
Our 'CASE WHEN <condition> ...' clauses should not contain query
parameters in the <condition> part, as this does not work with
server-side binding of parameters to query. This was previously abused
as psycopg2 does client-side bindings but it will not work for for
Psycopg 3.

Accordingly, we do this interpolation before passing the query string to
the database driver.