Use a fallback loader for character types if client encoding is ASCII
Despite our previous attempt to use the auto client encoding, we still
might get the 'ascii' encoding.

In that case, psycopg (3+) will not decode character type values, and
we'd get bytes when retrieving results:

  https://www.psycopg.org/psycopg3/docs/basic/adapt.html#strings-adaptation

Clearly, that's not what we expect and we thus need to decode things. To
do so, we register a loader for text and char types that simply decodes
(using client encoding, the real one, determined by Python) with a
replace policy on errors.