Use isort to automatically sort imports
https://pycqa.github.io/isort/ is nice to get imports ordered
automatically; so let's use it and configure it so that it plays nicely
with black.

Changes are mostly the result of running isort from the top-level
directory. Only part of changes in the 'pg' module are handled by hand,
typically the part between 'isort: off'/'isort: on' because isort
otherwise messes up 'type: ignore' comments.