Add new --queries-only flag for best viewport usage and performance
As there are quite many viewport customization flags, it's currently
quite tedious to set things for most compact viewport usage, when on
some terminal withing a terminal, as it quite often is on the clouds.

Other effect is that in this mode we're a bit lighter on the monitored
DB, skipping some queries.