Fix invalid call to pg_connect() from CLI main()
The second argument passed to pg_connect() is expected to be a
'exit_on_failed: bool', but we were passing the connection string. So when
the connection string is given by the user, it behaves similarly as
passing True (non-empty string evaluates to True); on the other hand, if
no connection string is passed, it evaluates to False, which disables
the 'exit_on_failed' feature. In fact, this exit_on_failed parameter is
never used by caller, so let's just remove it.