Add a few builtin profiles
We add a "minimal", a "narrow" and a "wide" built-in profiles, shipped
with pg_activity distribution. The first two hide non-necessary columns in
the process table and the first also hides all header information. The
"wide" profile is simply the inverse of "narrow", and may be used to
switch from one's custom profile to something close to the default
behavior.