Make headers info display configurable
In the command-line parser, options related to header information have
their default value set to None in order to distinguish when the option
was not specified from when it got, as it would store False then. This
makes it possible to have a precedence of command-line options over the
configuration file (implemented in types.UIHeader.make()).

The validation of configuration options, previously in UISection, is
extracted into a BaseSectionMixin and reused in new HeaderSection.