Make cell (default) color configurable
This only applies to columns with a fixed color, i.e. that does not
depend on the value (like do the 'time' or 'state' columns).

Although this limitation is not handled in the Configuration class, we
intercept invalid 'color' option in column configuration while loading
the types.UI object and still return a ConfigurationError so that the
user experience is quite similar.