Enable back setuptools automatic discovery
Automatic package discovery by setuptools was disabled in commit
8a5fa22373f, somehow as a workaround for an issue RPM build system
(https://github.com/rpm-software-management/rpm/issues/2532) or maybe
because it was misconfigured then.

However, having to declare packages explicitly (options 'packages' in
'[tool.setuptools]' section of pyproject.toml) is tedious and
error-prone as shown in https://github.com/dalibo/pg_activity/issues/411
which indicates that we were missing some sub-packages.

In the meantime, the RPM issue got resolved, so it seems safe to get
back to automatic discovery as this resolves the later issue in a more
future-proof way than explicitly listing packages.

So we get back to automatic discovery here, although not using
[tool.setuptools.packages.find], but simply letting setuptools discover
the "flat-layout" we're using, as can be seen by the following messages
from 'python -m build' un:

    * Building sdist...
    No `packages` or `py_modules` configuration, performing automatic discovery.
    `flat-layout` detected -- analysing .
    discovered packages -- ['pgactivity', 'pgactivity.profiles', 'pgactivity.queries']

This requires no configuration as we use common excluded names (like
'tests', 'docs').