Reverse the logic of --no-XXX option
We avoid propagating the negation and instead use a meaningful name as
"dest", set the action to "store_false" and the default to True.

Also the "dest" for "--no-database" is changed to "database" for
consistency with other options.

When building the Flag value, in Flag.load(), we now follow the order of
options definition.

Finally, we define a flag() helper to factor out the common definition
pattern of these options.